Comming after completing unit test methods for Premium Service methods

+++ PremiumCalc.API/Utility/IPremiumCalcLogic.cs
@@ -7,6 +7,6 @@ namespace PremiumCalc.API.Utility
 {
     public interface IPremiumCalcLogic
     {
-        decimal MonthlyPremiumCalcForUser(int DeathCoverAmt, int OccupationId, int Age);
+        double MonthlyPremiumCalcForUser(int DeathCoverAmt, int OccupationId, int Age);
     }
 }
